requirements.txt
absl-py==1.0.0
astunparse==1.6.3
biopython==1.79
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorhash==1.0.4
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
numpy==1.22.0
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
